+++ CHANGELOG.md
@@ -6,11 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- Datalog: pass sample to persistence layer in order
 
 ### Added
 
 - Improving errors returned by the datastore gRPC server.
+
+### Changed
+
+- Datalog: preserve order when passing samples to the persistence layer.
 
 ## v2.21.2 - 2023-12-04